# resources/openapi/tabou2-api.json
{
	"openapi": "3.0.3",
	"info": {
		"description": "You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
		"version": "1.0.0",
		"title": "tabou2",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"email": "apiteam@swagger.io"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"servers": [
		{
			"url": "/{basePath}",
			"description": "Tabou2 v2 API",
			"variables": {
				"basePath": {
					"default": "/v2"
				}
			}
		}
	],
...
	"paths": {
...
		"/operations/{operationId}/documents": {
			"get": {
				"summary": "Recherche des documents d'une opération",
				"operationId": "searchOperationDocuments",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant de l'opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "query",
						"name": "nom",
						"description": "Nom du document",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "libelleTypeDocument",
						"description": "Libellé du type de document",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "typeMime",
						"description": "Type MIME du document",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "start",
						"in": "query",
						"required": false,
						"description": "Index de début - démarre à 0",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "resultsNumber",
						"in": "query",
						"required": false,
						"description": "Nombre de résultats à remonter",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "orderBy",
						"in": "query",
						"required": false,
						"description": "Nom du champ sur lequel trier les résultats",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "asc",
						"in": "query",
						"required": false,
						"description": "Indique si le tri doit être fait par ordre ascendant",
						"schema": {
							"type": "boolean",
							"default": true
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/PageResult"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			},
			"post": {
				"summary": "Ajout d'un document à une opération",
				"operationId": "addOperationDocument",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant de l'opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "query",
						"name": "nom",
						"description": "Nom du document",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "libelleTypeDocument",
						"description": "Libellé du type de document",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "dateDocument",
						"description": "Date du document",
						"required": false,
						"schema": {
							"type": "string",
							"format": "date"
						}
					}
				],
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"fileToUpload": {
										"type": "string",
										"format": "binary",
										"description": "Document à ajouter"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/DocumentMetadata"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/operations/{operationId}/documents/{documentId}": {
			"get": {
				"summary": "Récupération des métadonnées d'un document d'une opération",
				"operationId": "getOperationDocumentMetadata",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant de l'opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "path",
						"name": "documentId",
						"description": "Identifiant du document",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/DocumentMetadata"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			},
			"put": {
				"summary": "Mise à jour des métadonnées d'un document d'une opération",
				"operationId": "updateOperationDocumentMetadata",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant d'une opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "path",
						"name": "documentId",
						"description": "Identifiant du document",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Metadonnées du document",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "tabou2-model.json#/components/schemas/DocumentMetadata"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/DocumentMetadata"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			},
			"delete": {
				"summary": "Suppression d'un document d'une opération",
				"operationId": "deleteOperationDocument",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant de l'opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "path",
						"name": "documentId",
						"description": "Identifiant du document",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/operations/{operationId}/documents/{documentId}/content": {
			"get": {
				"summary": "Récupération du contenu d'un document d'une opération",
				"operationId": "getOperationDocumentContent",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant de l'opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "path",
						"name": "documentId",
						"description": "Identifiant du document",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/octet-stream": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"404": {
						"description": "La ressource demandée n'existe pas"
					}
				}
			},
			"put": {
				"summary": "Mise à jour du contenu d'un document d'une opération",
				"operationId": "updateOperationDocumentContent",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant de l'opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "path",
						"name": "documentId",
						"description": "Identifiant du document",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"fileToUpload": {
										"type": "string",
										"format": "binary",
										"description": "Document à mettre à jour"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/DocumentMetadata"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/operations/emprises": {
			"get": {
				"summary": "Récupération de la liste des emprises des opérations non suivies dans Tabou",
				"operationId": "getAvailableOperationEmprises",
				"parameters": [
					{
						"in": "query",
						"name": "natureId",
						"description": "Identifiant de la nature de l'opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "query",
						"name": "estSecteur",
						"description": "True si la recherche porte sur les secteurs, false si on recherche les opérations",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "nom",
						"description": "Nom de l'opération",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "start",
						"in": "query",
						"required": false,
						"description": "Index de début - démarre à 0",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "resultsNumber",
						"in": "query",
						"required": false,
						"description": "Nombre de résultats à remonter",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "orderBy",
						"in": "query",
						"required": false,
						"description": "Nom du champ sur lequel trier les résultats",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "asc",
						"in": "query",
						"required": false,
						"description": "Indique si le tri doit être fait par ordre ascendant",
						"schema": {
							"type": "boolean",
							"default": true
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/PageResult"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/operations/{operationId}/etapes": {
			"get": {
				"summary": "Liste des étapes d'avancement possibles pour une opération d'aménagement",
				"operationId": "getEtapesByOperationId",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant de l'opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "tabou2-model.json#/components/schemas/Etape"
									}
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/operations/{operationId}/tiers": {
			"post": {
				"summary": "Association d'un tiers à une operation",
				"operationId": "associateTiersToOperation",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant d'une opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"requestBody": {
					"description": "Association entre le tiers et le type tiers",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "tabou2-model.json#/components/schemas/TiersTypeTiers"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/AssociationTiersTypeTiers"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			},
			"get": {
				"summary": "Recherche les tiers d'une opération",
				"operationId": "searchTiersByOperationId",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant d'une opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"name": "libelle",
						"in": "query",
						"required": false,
						"description": "Libellé du tiers",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "start",
						"in": "query",
						"required": false,
						"description": "Index de début - démarre à 0",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "resultsNumber",
						"in": "query",
						"required": false,
						"description": "Nombre de résultats à remonter",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "orderBy",
						"in": "query",
						"required": false,
						"description": "Nom du champ sur lequel trier les résultats",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "asc",
						"in": "query",
						"required": false,
						"description": "Indique si le tri doit être fait par ordre ascendant",
						"schema": {
							"type": "boolean",
							"default": true
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/PageResult"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/operations/{operationId}/tiers/{associationTiersId}": {
			"put": {
				"summary": "Modification d'un tiers d'une opération ",
				"operationId": "updateTiersByOperationId",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant d'une opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "path",
						"name": "associationTiersId",
						"description": "Identifiant du l'association tiers - type tiers - operation",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"requestBody": {
					"description": "Association entre le tiers, le type tiers",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "tabou2-model.json#/components/schemas/TiersTypeTiers"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/AssociationTiersTypeTiers"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			},
			"delete": {
				"summary": "Supprime un tiers d'une opération ",
				"operationId": "deleteTiersFromOperation",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant d'une opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "path",
						"name": "associationTiersId",
						"description": "Identifiant du l'association tiers - type tiers - operation",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/operations/{operationId}/evenements": {
			"get": {
				"summary": "Recherche la liste des évènements d'une opération d'aménagement",
				"operationId": "getEvenementsByOperationId",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant de l'opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					},
					{
						"in": "query",
						"name": "start",
						"description": "Index de début - démarre à 0",
						"required": false,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"in": "query",
						"name": "resultsNumber",
						"description": "Nombre de résultats à remonter",
						"required": false,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"description": "Nom du champ sur lequel trier les résultats",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "asc",
						"description": "Indique si le tri doit être fait par ordre ascendant",
						"required": false,
						"schema": {
							"type": "boolean",
							"default": true
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/PageResult"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			},
			"post": {
				"summary": "Ajout d'un évènement à une opération d'aménagement",
				"operationId": "addEvenementByOperationId",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant du programme",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"requestBody": {
					"description": "Nouvel évènement",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "tabou2-model.json#/components/schemas/Evenement"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/Evenement"
								}
							}
						}
					}
				}
			},
			"put": {
				"summary": "Modification d'un évènement d'une opération d'aménagement",
				"operationId": "updateEvenementByOperationId",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant de l'opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"requestBody": {
					"description": "Evènement à modifier",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "tabou2-model.json#/components/schemas/Evenement"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/Evenement"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
		"/operations/etapes": {
			"get": {
				"summary": "Liste des étapes par opérations",
				"description": "Ensemble d'étapes d'operations",
				"operationId": "searchOperationsEtapes",
				"parameters": [
					{
						"in": "query",
						"name": "code",
						"description": "Recherche par code",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "libelle",
						"description": "Recherche par nom de l'étapes (avec *)",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "start",
						"description": "Index de début - démarre à 0",
						"required": false,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"in": "query",
						"name": "resultsNumber",
						"description": "Nombre de résultats à remonter",
						"required": false,
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"description": "Nom du champ sur lequel trier les résultats",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "asc",
						"description": "Indique si le tri doit être fait par ordre ascendant",
						"required": false,
						"schema": {
							"type": "boolean",
							"default": true
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "tabou2-model.json#/components/schemas/PageResult"
								}
							}
						}
					}
				}
			}
		},
		"/operations/{operationId}/fiche-suivi": {
			"get": {
				"summary": "Récupération de la fiche de suivi d'une opération",
				"operationId": "downloadOperationFicheSuivi",
				"parameters": [
					{
						"in": "path",
						"name": "operationId",
						"description": "Identifiant de l'opération",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/octet-stream": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"404": {
						"description": "La resource demandée n'existe pas"
					}
				}
			}
		},
...
	}
}